init zookeeper connector.
